% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-checking.R
\name{checkTable}
\alias{checkTable}
\title{Summarise Dataframe with Custom Grouping}
\usage{
checkTable(df, numericVars, categoricalVars, outputPath)
}
\arguments{
\item{df}{The data frame to be summarized.}

\item{numericVars}{Numerical variables to be used in the executive summary.}

\item{categoricalVars}{Variables to be summarized in detail.}

\item{outputPath}{File path to save the output workbook. Just mention the folder,
the function will assign "checkTable-{df}.xlsx" as the file name to the workbook generated.}
}
\value{
None. The function saves the output workbook to the specified file path.
}
\description{
Creates a workbook with multiple summaries of a given data.frame, split
across multiple sheets. A variety of statistical summaries are provided
for user-specified numeric variables, and the data is filtered to provide
total counts for each user-specified categorical variable.
}
