% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customised-geographic-concordance.R
\name{customConcord}
\alias{customConcord}
\title{Applying Custom Concordances}
\usage{
customConcord(
  df,
  joinColumn,
  concordanceColumn,
  targetColumn,
  ConcordanceSource = NULL,
  uniqueMapping = TRUE
)
}
\arguments{
\item{df}{The dataframe to which the targetColumn is applied}

\item{joinColumn}{The name of the column in data used to join with concordanceColumn in the concordance.}

\item{concordanceColumn}{The name of the column in the concordance representing the original categorization in data.}

\item{targetColumn}{The name of the column in the concordance representing the new categorization to be applied to data.}

\item{ConcordanceSource}{Optional file path to a custom concordance CSV file or a dataframe containing concordance data.
If NULL, a pre-loaded master concordance is used.}

\item{uniqueMapping}{Boolean value indicating whether to remove duplicates.
If TRUE, only the first occurrence of each start_area is kept.}
}
\value{
A dataframe with the concordance applied.
}
\description{
This function aligns a dataframe with a specified concordance mapping, enabling the
transformation of existing data categories to a new categorization system. It supports
both a pre-loaded master concordance and custom user-provided concordances. The
function is particularly useful for re-mapping or re-classifying data, such as geographic
areas, but is also adaptable to other types of categorical data. All concordances till
2024 are already loaded in this package. If your concordanceColumn and targetColumn are
already present in this package, there is no need to provide concordanceFilePath. If its not present
and you'd like the concordances in this package to be up-to-date, please contact the maintainer
of this package (details present in the DESCRIPTION file).
}
\details{
Run 'colnames(MasterConcordance)' to check whether the intended concordances are present in the package.
}
