% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customised-data-sorting.R
\name{customSort}
\alias{customSort}
\title{Sorting dataframe with custom priorities}
\usage{
customSort(df, subtotals = NULL)
}
\arguments{
\item{df}{the dataframe to be sorted.}

\item{subtotals}{A list where one mentions the column variable name and the variable instance that needs to be pushed to the bottom of the dataframe}
}
\value{
A sorted dataframe based on the defined priorities.
}
\description{
Sorts a dataframe based on a custom order, and can handle additional custom
sorting order provided in `subtotals`
}
\details{
The sorting is done in the following manner -

1) NA occurences in dataframe are displayed

2) Standard values in the dataframe

3) A custom set of values (that can be given as a list of input by the
user), configured through the subtotals variable

4) Total stated

5) Total
}
