% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customised-geographic-concordance.R
\name{.onAttach}
\alias{.onAttach}
\title{MasterConcordance Data Set}
\format{
A data frame with all concordances
}
\source{
database
}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{sample description for libname}

\item{pkgname}{sample description for pkgname}
}
\description{
A dataset containing the master concordance information.
This data set is loaded automatically in the user session when the package is loaded.
}
