% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-checking.R
\name{grrCheck}
\alias{grrCheck}
\title{Checking Graduated Random Rounding is applied to the variable or not
Performs a graduated random rounding check on column variable.
Rounds numbers based on defined ranges and checks for multiples.}
\usage{
grrCheck(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Logical value indicating if the numbers in `x` follow
  the graduated random rounding rules.
}
\description{
Checking Graduated Random Rounding is applied to the variable or not
Performs a graduated random rounding check on column variable.
Rounds numbers based on defined ranges and checks for multiples.
}
