% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-checking.R
\name{rr3Check}
\alias{rr3Check}
\title{Checking RandomRounding3 is applied to variable or not
Checks if all variables in a numeric vector are multiples of 3.}
\usage{
rr3Check(x)
}
\arguments{
\item{x}{column variable that needs checking}
}
\value{
Boolean indicating whether all values in `x` are multiples of 3.
}
\description{
Checking RandomRounding3 is applied to variable or not
Checks if all variables in a numeric vector are multiples of 3.
}
