% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-checking.R
\name{summariseAndFilter}
\alias{summariseAndFilter}
\title{Summarise and Filter Data}
\usage{
summariseAndFilter(inputDf, groupVar, columnOrder, detailSummaryVars)
}
\arguments{
\item{inputDf}{Data frame to be processed.}

\item{groupVar}{The variable to group by.}

\item{columnOrder}{Order of columns to maintain in the output.}

\item{detailSummaryVars}{Variables used for cumulative filtering.}
}
\value{
A summarized data frame after applying the filtering process.
}
\description{
Applies a cumulative filtering process on a data frame and then summarizes the data.
Specifically filters rows with a "Total" value in specified variables.
}
